refactor: rename file `utils.ts` <- `_utils.ts`
